+++ service/controller/aws/v15/resource/certconfigcleansing/delete.go
@@ -1,0 +1,7 @@
+package certconfigcleansing
+
+import "context"
+
+func (r Resource) EnsureDeleted(ctx context.Context, obj interface{}) error {
+	return nil
+}
